Renamed files to snake_case.
