+++ engine/Components/Verifier.py
@@ -1,4 +1,4 @@
 # todo Check connectives
 # todo Suggest list of connectives
-# todo From Sanitizer.py, check code and if not, return to user input or ErrorHandler.py
-# todo If Verifier.py is valid, proceed
+# todo From sanitizer.py, check code and if not, return to user input or ErrorHandler.py
+# todo If verifier.py is valid, proceed
