+++ engine/Computation/evaluate_statement.py
@@ -1,7 +1,7 @@
 from math import log2
 
-from Computation.SwitchConnectives import switch_case
-from Computation.TableMatrix.CreateTableMatrix import create_matrix
+from Computation.switch_connectives import switch_case
+from Computation.TableMatrix.create_table_matrix import create_matrix
 from Logging.logging_config import logger
