+++ engine/Simulator/input_statements.py
@@ -1,4 +1,4 @@
-# Test cases that is used in SimulateMain.py
+# Test cases that is used in simulate_main.py
 test_cases = [
 	# "(P <-> (P <-> (P <-> (P <-> P))))",
 	# "(P ^ Q)",  # valid
